+++ app/Http/Controllers/Auth/VerificationController.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace App\Http\Controllers\Auth;
+namespace App\Http\Controllers\Api\Api\Api\Auth;
 
-use App\Http\Controllers\Controller;
+use App\Http\Controllers\Api\Api\Api\Controller;
 use App\Providers\RouteServiceProvider;
 use Illuminate\Foundation\Auth\VerifiesEmails;
