+++ app/Http/Resources/Api/WedstrijdResource.php
@@ -1,8 +1,7 @@
 <?php
 
-namespace App\Http\Resources;
+namespace App\Http\Resources\Api;
 
-use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\JsonResource;
 
 class WedstrijdResource extends JsonResource
